+++ libs/angular/src/v-angular/base-control-value-accessor/base-control-value-accessor.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from '@angular/core'
+import { NgvBaseControlValueAccessorComponent } from './base-control-value-accessor.component'
+
+@NgModule({
+  declarations: [NgvBaseControlValueAccessorComponent],
+  imports: [],
+  exports: [NgvBaseControlValueAccessorComponent],
+})
+export class NgvBaseControlValueAccessorModule {}