+++ libs/angular/src/v-angular/dropdown/typeahead/index.ts
@@ -1,0 +1,5 @@
+export * from './typeahead-highlight/typeahead-highlight.component'
+export * from './typeahead-input/typeahead-input.component'
+export * from './typeahead-dropdown-list/typeahead-dropdown-list.component'
+export * from './typeahead.directive'
+export * from './typeahead.module'