+++ libs/angular/src/v-angular/textarea/textarea.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core'
+import { CommonModule } from '@angular/common'
+
+import { NgvTextareaComponent } from './textarea.component'
+
+@NgModule({
+  declarations: [NgvTextareaComponent],
+  imports: [CommonModule],
+  exports: [NgvTextareaComponent],
+})
+export class NgvTextareaModule {}