+++ libs/core/src/components/button/fab/fab.ts
@@ -22,6 +22,7 @@ export class GdsFab extends GdsButton {
    */
   @styleExpressionProperty({
     valueTemplate: (v) => v,
+    selector: '[part="_button"]',
   })
   position?: string