+++ libs/core/src/utils/helpers/style-expression-parser.ts
@@ -1,6 +1,9 @@
+import { tr } from 'date-fns/locale'
+
 type Tokens = string[]
 type BreakpointSpecifier = { condition: string; value: string }[]
-type BreakpointData = { breakpoint: string; values: string[] }
+type BreakpointValues = { sel: string; values: string[] }
+type BreakpointData = { breakpoint: string; values: BreakpointValues[] }
 type BreakpointTree = BreakpointData[]
 
 const viewportBreakpoints: Record<string, string> = {
@@ -20,7 +23,7 @@ const viewportBreakpoints: Record<string, string> = {
 
 const breakpointValueRegex = /^([<|>]=?)?([0-9a-z]+)/
 
-const singleCharTokens = ['{', '}', ';', ',']
+const controlTokens = ['{', '}', ';', ':', ',']
 const whitespace = [' ', '/n']
 
 /**
@@ -38,7 +41,7 @@ export function tokenize(source: string): Tokens {
 
     if (!whitespace.includes(c)) scanned += c
 
-    if (singleCharTokens.includes(c)) {
+    if (controlTokens.includes(c)) {
       lexemes.push(scanned.slice(0, -1))
       lexemes.push(c)
       scanned = ''
@@ -62,60 +65,70 @@ export function tokenize(source: string): Tokens {
  * @returns A tree structure
  */
 export function parse(tokens: Tokens): BreakpointTree {
-  const tree = []
-  let level = 'bp'
-  let scope: BreakpointData | undefined = undefined
-  let expectMultiCondition = false
+  // This is the tree we are building and returning at the end
+  const tree: BreakpointTree = []
 
+  // The current breakpoint scope we are adding to
+  let bpScope: BreakpointData = { breakpoint: '-', values: [] }
+
+  // A function to create an empty breakpoint value bucket
+  const getEmptyBucket = (): BreakpointValues => ({ sel: '', values: [] })
+
+  // The current value bucket we are adding to
+  let valueBucket: BreakpointValues = getEmptyBucket()
+
+  // Step through the tokens
   for (const t of tokens) {
-    if (!singleCharTokens.includes(t)) {
-      if (level === 'val' && scope) {
-        t !== '}' && scope.values.push(t)
-      } else {
-        if (scope && expectMultiCondition) {
-          scope.breakpoint += `,${t}`
-          expectMultiCondition = false
-          continue
-        }
-        if (!scope) {
-          scope = { breakpoint: t, values: [] }
-          tree.push(scope)
-        } else {
-          // This means short hand 0 breakpoint
-          level = 'val'
-          scope.values.push(scope.breakpoint)
-          scope.values.push(t)
-          scope.breakpoint = '-'
-        }
-      }
+    // Add non-control tokens to the value bucket
+    if (!controlTokens.includes(t)) {
+      valueBucket.values.push(t)
       continue
     }
 
-    if (t === ',') {
-      expectMultiCondition = true
-      continue
-    }
-
+    // If we come across '{', then create a new breakpoint scope, and empty the
+    // value bucket into the breakpoint specifier
     if (t === '{') {
-      level = 'val'
-      continue
+      bpScope = { breakpoint: valueBucket.values.join(','), values: [] }
+      valueBucket = getEmptyBucket()
     }
 
-    if (t === '}' || t === ';') {
-      level = 'bp'
-      scope = undefined
-      continue
+    // If we come across ';', then empty the value bucket into the breakpoint scope
+    // And if the tree is still empty at this point, we add the breakpoint scope to the tree
+    if (t === ';') {
+      if (tree.length === 0) {
+        tree.push(bpScope)
+      }
+      if (valueBucket.values.length > 0) {
+        bpScope.values.push(valueBucket)
+        valueBucket = getEmptyBucket()
+      }
+    }
+
+    // If we come across ':', then pop the last value from the value bucket and use
+    // it as the selector for a new value bucket
+    if (t === ':') {
+      const sel = valueBucket.values.pop() ?? ''
+      valueBucket.sel = sel
+    }
+
+    // If we come across '}', then close the breakpoint scope and add it to the tree
+    if (bpScope && t === '}') {
+      bpScope.values.push(valueBucket)
+      valueBucket = getEmptyBucket()
+      tree.push(bpScope)
     }
   }
 
-  return tree.map((bp) => {
-    // Fix remaining shorthand 0 breakpoints
-    if (bp.values.length === 0) {
-      bp.values.push(bp.breakpoint)
-      bp.breakpoint = '-'
-    }
-    return bp
-  })
+  // If the value bucket is not empty at the end, add it to a new '-' breakpoint
+  if (valueBucket.values.length > 0) {
+    bpScope.values.push(valueBucket)
+  }
+
+  if (tree.length === 0) {
+    tree.push(bpScope)
+  }
+
+  return tree
 }
 
 function parseBreakpoint(bp: string): BreakpointSpecifier {
@@ -152,7 +165,30 @@ export function toCss(
           `(${b.condition?.includes('<') ? 'max-width' : 'min-width'}: ${viewportBreakpoints[b.value] ?? b.value})`,
       )
       .join(' and ')
-    const mq = `@media ${query} {${selector}{${styleTemplate(property, bp.values.map(valueTemplate))}}}`
+
+    // For each set of values in the breakpoint, construct a selector based on the
+    // specified selector and the base selector, and apply the style template to each
+    const mq = `@media ${query} {${bp.values
+      .map((bpValues: BreakpointValues) => {
+        let sel = selector
+        if (bpValues.sel.length > 0)
+          sel =
+            selector === ':host'
+              ? `:host(:${bpValues.sel})`
+              : `${selector}:${bpValues.sel}`
+
+        const style = styleTemplate(
+          property,
+          bpValues.values.map(valueTemplate),
+        )
+
+        // If the selector is hover, we wrap the style in a hover media query so that
+        // it excludes touch devices
+        if (bpValues.sel === 'hover')
+          return `@media (hover: hover) {${sel}{${style}}}`
+        else return `${sel}{${style}}`
+      })
+      .join('')}}`
     css += mq
   }